placing the source files
This FFT implementation is a personal implementation of the Goertzel algorithm. It is intended to be used in synergy with the Polyphonic Pith Detection algorithm which uses the RTFI resonator technique. The FFT should be used in order to get the correct amplitude of the peaks because the RTFI filters take a certain a mount of time to stabilize, this means that if you need the exact amplitude quickly they won't be suitable, they are, instead, very useful for discriminating between close peaks.